requirements.txt
Django==1.7.6
Pillow==2.7.0
South==1.0.2
argparse==1.2.1
crispy-forms-foundation==0.4.1
django-ajax-selects==1.3.6
django-analytical==0.21.0
django-appconf==1.0.1
django-crispy-forms==1.4.0
django-disqus==0.5
django-easy-select2==1.2.12
django-endless-pagination==2.0
django-froala-editor==1.2.6
django-guardian==1.2.5
django-infinite-pagination==1.0
django-pure-pagination==0.2.1
django-registration-redux==1.1
django-user-accounts==1.0.1
django-userena==1.4.0
easy-thumbnails==2.2
gunicorn==19.3.0
html2text==2015.2.18
mock==1.0.1
phileo==1.3
pytz==2015.2
sass==2.3
six==1.9.0
wsgiref==0.1.2
